% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning the NOAA Earthquake Dataset}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{A raw data frame obtained from NOAA}
}
\value{
A cleaned data frame with \code{date} is created by uniting \code{YEAR},
\code{MONTH}, \code{DAY}, \code{LONGITUDE} and \code{LATITUDE} are converted to numeric
}
\description{
\code{eq_clean_data()} takes the raw NOAA data frame and returns a clean data frame,
which would be well-behaved with visualise functions
in this package.
}
\note{
The dataset is included in the package and could be called by
\code{data(raw_data)}. Observations with \code{NA} \code{DAY} or/and \code{MONTH}
would take the value of \code{01}. The dataset includes dates BEC (with negative years).
}
\examples{
\dontrun{use("raw_data")
data <- eq_clean_data(raw_data)}

}
