% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_country_filter}
\alias{eq_country_filter}
\title{Filter the NOAA dataset by countries and time range}
\usage{
eq_country_filter(data, countries, xmin, xmax)
}
\arguments{
\item{data}{A data frame with relevant columns}

\item{countries}{A character vectors with the names of considered countries}

\item{xmin}{The beginning of considered timeline, as a date object}

\item{xmax}{The ending of considered timeline, as a date object}
}
\value{
A subset of the inputed data frame
}
\description{
\code{eq_country_filter()} returns
the data specifically for called \code{countries},
in the timeline from \code{xmin} to \code{xmax}
}
\note{
The \code{xmin} and \code{xmax} must be date object
(see: \code{lubridate::ymd()})
}
\examples{
\dontrun{input_data <- eq_country_filter(data,
c("INDIA", "THAILAND"),
xmin = as.Date("1800-01-01"),
xmax = as.Date("2010-01-01"))}
}
