% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creating the pop-up annotation}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{The inputed data frame}
}
\description{
Function \code{eq_create_label()} takes the dataset as an argument and creates an HTML label
that can be used as the annotation text in the leaflet map.
}
\examples{
\dontrun{
eq_country_filter(data, c("INDIA"),
xmin = as.Date("1850-01-01"),
xmax = as.Date("1920-01-01"))\%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}
}
