% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning the LOCATION_NAME}
\usage{
eq_location_clean(raw_data)
}
\arguments{
\item{raw_data}{A data frame with uncleaned \code{LOCATION_NAME}}
}
\value{
A data frame with cleaned \code{LOCATION_NAME}
}
\description{
\code{eq_location_clean()} cleans \code{LOCATION_NAME} column in the raw NOAA data frame
by returning only the names of locations without \code{COUNTRY} (avoid redundant) in
the formate of title case, which could be used as pop-up text in leaflet.
}
\examples{
\dontrun{data <- eq_location_clean(eq_clean_data(raw_data))}

}
