% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{Mapping Tools for the NOAA Earthquake Dataset}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A filtered data frame with earthquakes to visualize}

\item{annot_col}{The name of column is used for the annotation}
}
\description{
Function \code{eq_map()} takes an argument data containing the filtered
data frame with earthquakes to visualise in the geographical maps bu \code{LONGITUDE}
and \code{LATITUDE}, with pop-up windows to provide the further information.
}
\examples{
\dontrun{
eq_country_filter(data, c("INDIA"),
xmin = as.Date("1850-01-01"),
xmax = as.Date("1920-01-01")) \%>\%
eq_map(annot_col = "date")}

}
