% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{NOAA Significant Earthquakes Dataset}
\format{A data frame with 5958 obs. of 47 variables:}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
data("raw_data")
}
\description{
This dataset contains information about 5,933 earthquakes
over an approximately 4,000 year time span.
}
\keyword{datasets}
