% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_stat_timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme of Timeline}
\usage{
theme_timeline()
}
\description{
\code{theme_timeline()} is used with \code{geom_timeline()}
and \code{geom_timeline_label()}
}
\note{
The background and y axis is set as blank, when only the line axis x
is presented
}
\examples{
\dontrun{
ggplot(input_data, aes (x = date, y = COUNTRY,
color = as.numeric(DEATHS), size = as.numeric(EQ_PRIMARY))) +
geom_timeline(aes(xmin = as.Date("1900-01-01"), xmax =  as.Date("1925-01-01")) +
geom_timeline_label(aes(label = LOCATION_NAME,
xmin = as.Date("1900-01-01"), xmax = as.Date("1925-01-01")),
n_max = 5) +
theme_timeline()}
}
